mdx bundler complete
