+++ lib/xdm.js
@@ -16,10 +16,6 @@ import rehypeSlug from 'rehype-slug';
 import rehypeKatex from 'rehype-katex';
 import rehypePrism from '@mapbox/rehype-prism';
 
-import { join } from 'path';
-import { readdirSync, readFileSync } from 'fs';
-import { bundleMDX } from 'mdx-bundler';
-
 export const useXdm = (defaults) => {
   const [state, setState] = useState({ ...defaults, file: null });
   const { run: setConfig } = useDebounceFn(
@@ -65,42 +61,4 @@ export const useXdm = (defaults) => {
   );
 
   return [state, setConfig];
-};
-
-export const getAllSlugs = async (directory) => {
-  const postsPath = join(process.cwd(), 'posts', directory);
-  const fileNames = readdirSync(postsPath);
-
-  return fileNames;
-};
-
-export const getPostBySlug = async (directory, slug) => {
-  if (process.platform === 'win32') {
-    process.env.ESBUILD_BINARY_PATH = join(
-      process.cwd(),
-      'node_modules',
-      'esbuild',
-      'esbuild.exe'
-    );
-  } else {
-    process.env.ESBUILD_BINARY_PATH = join(
-      process.cwd(),
-      'node_modules',
-      'esbuild',
-      'bin',
-      'esbuild'
-    );
-  }
-
-  const mdxSource = readFileSync(
-    join(process.cwd(), 'posts', directory, `${slug}.mdx`),
-    'utf-8'
-  ).trim();
-
-  const { code, frontmatter } = await bundleMDX(mdxSource);
-
-  return {
-    code,
-    frontmatter,
-  };
 };
